test(fixtures): relocate sanitize changelog functions
